refactor: moves project files into src/lib
