+++ package.json
@@ -58,6 +58,7 @@
     "@semantic-release/release-notes-generator": "^14.0.3",
     "@types/jest": "^29.5.14",
     "@types/node": "^22.13.4",
+    "@types/stats.js": "^0.17.3",
     "@types/three": "^0.173.0",
     "@typescript-eslint/eslint-plugin": "^8.24.0",
     "dts-bundle-generator": "^9.5.1",
