+++ src/lib/services/instancedmesh/instancedMeshManager.ts
