+++ src/lib/services/instancedmesh/shaders/instanceFragmentShader.ts
