+++ src/lib/services/instancedmesh/shaders/instanceVertexShader.ts
